+++ src/core/enhancement/AdtEnhancement.ts
@@ -153,7 +153,8 @@ export class AdtEnhancement
           badi_definition: config.badiDefinition,
           masterSystem: this.systemContext.masterSystem,
           responsible: this.systemContext.responsible,
-          masterLanguage: this.systemContext.masterLanguage,
+          masterLanguage:
+            config.masterLanguage ?? this.systemContext.masterLanguage,
         },
         this.logger,
       );