+++ src/core/program/AdtProgram.ts
@@ -141,7 +141,8 @@ export class AdtProgram implements IAdtObject<IProgramConfig, IProgramState> {
           sourceCode: options?.sourceCode || config.sourceCode,
           masterSystem: this.systemContext.masterSystem,
           responsible: this.systemContext.responsible,
-          masterLanguage: this.systemContext.masterLanguage,
+          masterLanguage:
+            config.masterLanguage ?? this.systemContext.masterLanguage,
         },
         this.contentTypes,
       );