+++ src/core/structure/AdtStructure.ts
@@ -122,7 +122,8 @@ export class AdtStructure
         description: config.description,
         masterSystem: this.systemContext.masterSystem,
         responsible: this.systemContext.responsible,
-        masterLanguage: this.systemContext.masterLanguage,
+        masterLanguage:
+          config.masterLanguage ?? this.systemContext.masterLanguage,
       });
       objectCreated = true;
       state.createResult = createResponse;