+++ src/core/table/AdtTable.ts
@@ -100,7 +100,8 @@ export class AdtTable implements IAdtObject<ITableConfig, ITableState> {
         ddl_code: options?.sourceCode || config.ddlCode,
         masterSystem: this.systemContext.masterSystem,
         responsible: this.systemContext.responsible,
-        masterLanguage: this.systemContext.masterLanguage,
+        masterLanguage:
+          config.masterLanguage ?? this.systemContext.masterLanguage,
       });
       objectCreated = true;
       state.createResult = createResponse;