+++ src/core/tabletype/AdtDdicTableType.ts
@@ -103,7 +103,8 @@ export class AdtDdicTableType
         transport_request: config.transportRequest,
         masterSystem: this.systemContext.masterSystem,
         responsible: this.systemContext.responsible,
-        masterLanguage: this.systemContext.masterLanguage,
+        masterLanguage:
+          config.masterLanguage ?? this.systemContext.masterLanguage,
       });
       objectCreated = true;
       state.createResult = createResponse;