+++ src/core/transformation/AdtTransformation.ts
@@ -138,7 +138,8 @@ export class AdtTransformation
         description: config.description,
         masterSystem: this.systemContext.masterSystem,
         responsible: this.systemContext.responsible,
-        masterLanguage: this.systemContext.masterLanguage,
+        masterLanguage:
+          config.masterLanguage ?? this.systemContext.masterLanguage,
       });
       state.createResult = createResponse;
       this.logger?.info?.('Transformation created');